# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Evermos service backend.",
        "title": "Evermos Store and Product API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "@habbazettt"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login a user and return a JWT token along with user details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login a user with phone number and password",
                "parameters": [
                    {
                        "description": "Login Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user with the provided details (name, phone number, email, password, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user to the system",
                "parameters": [
                    {
                        "description": "Register Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.LoginRequest": {
            "type": "object",
            "properties": {
                "kata_sandi": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                }
            }
        },
        "controllers.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id_kota": {
                    "$ref": "#/definitions/services.City"
                },
                "id_provinsi": {
                    "$ref": "#/definitions/services.Province"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "pekerjaan": {
                    "type": "string"
                },
                "tanggal_Lahir": {
                    "type": "string"
                },
                "tentang": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "controllers.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id_kota": {
                    "type": "string"
                },
                "id_provinsi": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "jenis_kelamin": {
                    "type": "string"
                },
                "kata_sandi": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "pekerjaan": {
                    "type": "string"
                },
                "tanggal_lahir": {
                    "type": "string"
                },
                "tentang": {
                    "type": "string"
                }
            }
        },
        "controllers.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "services.City": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "province_id": {
                    "type": "string"
                }
            }
        },
        "services.Province": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}